Restore Index Preview (1.6)
I restored the color preview in the entries table that somehow
disappeared in the latest versions of this extension.

Made the preview a little smaller (size of a checkbox) than it used to
be - suits the ‘data’-oriented look of the table better in my eyes.

Did also switch to using delegates for the assets and tweaked the
settings panel a little bit.

I packed this up as version 1.6 that’s ready to release if you’re ok
with these changes.